templates/requirements.txt
amqp==1.0.13
anyjson==0.3.3
billiard==2.7.3.31
boto==2.34.0
celery==3.0.20
celery-with-redis==3.0
defusedxml==0.4.1
Django==1.5.11
django-extensions==1.1.1
django-pagination==1.0.7
django-pagination-bootstrap==1.0.7
django-s3-cache==1.3
django-storages==1.1.8
django-tastypie==0.9.14
djangorestframework==2.3.13
djangotoolbox==1.6.2
gunicorn==0.17.4
kombu==2.5.13
meld3==0.6.10
mimeparse==0.1.3
mock==1.0.1
phpserialize==1.3
Pygments==1.6
pymongo==2.8
PyMySQL==0.7.11
python-dateutil==2.1
python-mimeparse==0.1.4
redis==2.7.6
requests==1.1.0
simplejson==3.11.1
six==1.3.0
South==1.0
supervisor==3.0
virtualenv==13.0.3

